requirements.txt
openai-whisper
fuzzywuzzy
python-Levenshtein
